Merging refactored client in

+++ src/sdk/Townsharp.Infrastructure/Identity/BotTokenProvider.cs
@@ -34,7 +34,6 @@ public class BotTokenProvider : IBotTokenProvider
 
         FormUrlEncodedContent content = new FormUrlEncodedContent(request);
 
-        // This is not (yet) fault tolerant, but it needs to be, so we could use Polly for this, but if we are going to keep the code free of dependencies, we need to write our own retry logic.
         this.tokenCache = new AsyncCache<string>(
             TimeSpan.FromMinutes(15),
             async (CancellationToken cancellationToken) =>
@@ -62,13 +61,13 @@ public class BotTokenProvider : IBotTokenProvider
         return this.tokenCache.GetAsync(cancellationToken);
     }
 
-    public async ValueTask<ulong> GetBotUserIdAsync(CancellationToken cancellationToken = default)
+    public async ValueTask<int> GetBotUserIdAsync(CancellationToken cancellationToken = default)
     {
         var token = await this.GetTokenAsync(cancellationToken).ConfigureAwait(false);
         var claims = JwtDecoder.DecodeJwtClaims(token);
 
         string userIdString = claims["client_sub"]?.GetValue<string>() ?? throw new InvalidOperationException("Unable to find client_sub claim on JWT, so not bot user id could be determined.");
-        return UInt64.Parse(userIdString);
+        return int.Parse(userIdString);
     }
 
     public async ValueTask<string> GetBotUserNameAsync(CancellationToken cancellationToken = default)
