+++ src/sdk/Townsharp.Infrastructure/Subscriptions/SubscriptionDefinition.cs
@@ -1,6 +1,6 @@
 ﻿namespace Townsharp.Infrastructure.Subscriptions;
 
-public record SubscriptionDefinition(string EventId, long KeyId)
+public record SubscriptionDefinition(string EventId, int KeyId)
 {
     public static SubscriptionDefinition Parse(string eventPath)
     {
@@ -10,7 +10,7 @@ public record SubscriptionDefinition(string EventId, long KeyId)
             throw new ArgumentException("Invalid subscription definition format.  Expected <event>/<key>", nameof(eventPath));
         }
 
-        return new SubscriptionDefinition(parts[0], long.Parse(parts[1]));
+        return new SubscriptionDefinition(parts[0], int.Parse(parts[1]));
     }
 
     override public string ToString() => $"{EventId}/{KeyId}";
